fix(XkcdResponse): fix `link` field deserialization
Change the type of the `link` field from `Url` to `String`. This is necessary
because `Url` does not deserialize "" properly. Replacing `link` with the type
`Option<Url>` was also not feasible as that does not fix the issue.

Added test for deserializing `XkcdResponse` strings with a blank `link` field.

BREAKING CHANGE: `link` field type was changed from `Url` to `String`.